[None][refactor] Rename sparse_forward.py to flash_mla.py in dsa/
Better reflects the file's content: FlashMLA sparse attention kernel
path and short-seq MHA optimization helper.
